fix corrections requested
