+++ package.json
@@ -23,6 +23,7 @@
     "@lucia-auth/adapter-prisma": "^4.0.1",
     "@next/third-parties": "^14.2.15",
     "@radix-ui/react-accordion": "^1.2.1",
+    "@radix-ui/react-alert-dialog": "^1.1.2",
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-checkbox": "^1.1.2",
     "@radix-ui/react-dialog": "^1.1.2",
